+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Remove sign in message for voting
 * Move add suggestion row to top of search results
 * Move email/name fields below category on new suggestion view
+* Back out to the welcome screen after creating a suggestion
 
 ## 1.0.0 ##